Miscellaneous Update
This commit includes various changes to the repo. There is now a test
directory with a basic unit test case that can be run through Gradle.
The Network, Centrality, and DegreeCentrality classes have been updated
to support Number subclasses, rather than just the Integer class. This
change was necessary to allow for centrality metrics that return float
values for centrality. Finally, the Network class has an added
constructor to support Erdos-Renyi random networks.