Update (#16)
* Project Restructure

Note: Code is *not* currently functional and will not be compiled if
you try to run it at this point.

This commit simply serves as a point that provides a clear direction as
to how to handle Simulation setup. Methods and stepwise refinement
needs to be done here to make the code more maintainable and to a point
where it can run successfully.

Also, project has been restructured. Prior, the project was very messy
with all classes being in the same package. Now, classes that are
tightly coupled belong in their own package.

* Cleaned Up Simulations

Simulation code is significantly cleaner. Main class now has a working
simulation sample.

TODO: Implement simulations for interdependent networks (IDN).

* IDN Simulations

Implemented IDN simulations, with the ability to specify whether
bridging occurs before or after centrality determination.

* Modified Main